requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
supabase==2.7.4
python-dotenv==1.0.1
openai==1.46.0
sendgrid==6.11.0
pydantic==2.9.1
email-validator==2.2.0
httpx==0.27.0
requests==2.32.3
langsmith==0.1.131
pytest
